refactor: standardize build output directory from 'dist' to 'build' across monorepo
- Update all tsconfig.json files to use "outDir": "./build"
- Remove 'dist' entries from .gitignore files (build already present)
- Update package.json "main" fields from "dist/index.js" to "build/index.js"
- Update all setup-dev.js and prepare-publish.js scripts to reference build directories
- Update import paths in test files from test-mcp-client/dist to test-mcp-client/build
- Clean up ESLint and Vitest configs to remove redundant dist references
- Add root clean script using glob patterns for experimental/**/build and productionized/**/build

This change standardizes the build output directory naming convention across all MCP servers
for better consistency and simpler clean operations.
