+++ .eslintrc.json
@@ -17,5 +17,5 @@
     "@typescript-eslint/no-unused-vars": ["error", { "argsIgnorePattern": "^_" }],
     "@typescript-eslint/no-explicit-any": "error"
   },
-  "ignorePatterns": ["node_modules", "dist", "build", "*.js", "*.mjs", "*.cjs", "**/local/shared"]
+  "ignorePatterns": ["node_modules", "build", "*.js", "*.mjs", "*.cjs", "**/local/shared"]
 }
