+++ experimental/appsignal/tests/manual/performance-tools.manual.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeAll, afterAll } from 'vitest';
-import { TestMCPClient } from '../../../../test-mcp-client/dist/index.js';
+import { TestMCPClient } from '../../../../test-mcp-client/build/index.js';
 import path from 'path';
 import { fileURLToPath } from 'url';
 import type { TimelineEvent } from '../../shared/src/appsignal-client/lib/performance-incident-sample-timeline.js';
