+++ experimental/appsignal/tsconfig.json
@@ -10,5 +10,5 @@
     "noEmit": true
   },
   "include": ["tests/**/*.ts"],
-  "exclude": ["node_modules", "**/dist", "**/build"]
+  "exclude": ["node_modules", "**/build"]
 }
