+++ experimental/twist/.gitignore
@@ -3,7 +3,6 @@ node_modules/
 
 # Build output
 build/
-dist/
 *.tsbuildinfo
 
 # Environment files
