+++ experimental/twist/package.json
@@ -14,7 +14,7 @@
     "ci:install": "npm ci && cd shared && npm ci && cd ../local && npm ci",
     "build": "cd shared && npm run build && cd ../local && npm run build",
     "build:test": "cd shared && npm run build && cd ../local && npm run build && cd ../../../test-mcp-client && npm run build",
-    "clean": "find . -name '*.js' -not -path './node_modules/*' -not -path './*/node_modules/*' -not -path './local/build/*' -not -path './shared/dist/*' -not -path './*/dist/*' -not -path './*/build/*' -delete && find . -name '*.js.map' -not -path './node_modules/*' -not -path './*/node_modules/*' -not -path './local/build/*' -not -path './shared/dist/*' -not -path './*/dist/*' -not -path './*/build/*' -delete",
+    "clean": "find . -name '*.js' -not -path './node_modules/*' -not -path './*/node_modules/*' -not -path './local/build/*' -not -path './shared/build/*' -not -path './*/build/*' -delete && find . -name '*.js.map' -not -path './node_modules/*' -not -path './*/node_modules/*' -not -path './local/build/*' -not -path './shared/build/*' -not -path './*/build/*' -delete",
     "dev": "cd local && npm run dev",
     "test": "vitest",
     "test:run": "vitest run",
