+++ experimental/twist/shared/package.json
@@ -2,7 +2,7 @@
   "name": "twist-mcp-server-shared",
   "version": "0.1.1",
   "description": "Shared code for Twist MCP Server implementations",
-  "main": "dist/index.js",
+  "main": "build/index.js",
   "type": "module",
   "scripts": {
     "build": "tsc",
