+++ experimental/twist/shared/tsconfig.json
@@ -8,11 +8,11 @@
     "skipLibCheck": true,
     "declaration": true,
     "declarationMap": true,
-    "outDir": "./dist",
+    "outDir": "./build",
     "rootDir": "./src",
     "lib": ["es2022"],
     "types": ["node"]
   },
   "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"]
+  "exclude": ["node_modules", "build"]
 }
