+++ experimental/twist/tests/integration/twist.integration.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach, afterEach } from 'vitest';
-import { TestMCPClient } from '../../../../test-mcp-client/dist/index.js';
+import { TestMCPClient } from '../../../../test-mcp-client/build/index.js';
 // import { createIntegrationMockTwistClient } from '../../shared/src/twist-client/twist-client.integration-mock.js';
 import path from 'path';
 import { fileURLToPath } from 'url';
