+++ experimental/twist/tests/manual/test-robust-pagination.ts
@@ -1,5 +1,5 @@
 import { config } from 'dotenv';
-import { TestMCPClient } from '../../../../test-mcp-client/dist/index.js';
+import { TestMCPClient } from '../../../../test-mcp-client/build/index.js';
 import path from 'path';
 import { fileURLToPath } from 'url';
