+++ mcp-server-template/.gitignore
@@ -1,6 +1,5 @@
 node_modules/
 build/
-dist/
 .env
 .env.local
 .DS_Store
